feat(Offer): adaptação de Offer para necessitar de uma FlowerShop no construtor

+++ SeedShareWS/src/main/java/com/seedshare/service/offer/OfferServiceImpl.java
@@ -24,8 +24,9 @@ import com.seedshare.repository.StateRepository;
 import com.seedshare.repository.UserRepository;
 
 /**
- * Implementation of Offer Service interface
+ * Implementation of {@link com.seedshare.service.offer.OfferService} interface
  * 
+ * @author gabriel.schneider
  * @author joao.silva
  */
 @Service
@@ -56,10 +57,10 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 	public ResponseEntity<?> save(Offer offer) {
 		if(isNotNull(offer)) {
 			Offer newOffer = new Offer(offer.getUnitPrice(), offer.getAmount(), getCurrentUser(), offer.getSpecies(),
-					offer.getDescription());
+					offer.getDescription(), getCurrentUser().getFlowerShop());
 			if (newOffer.isValid()) {
 				newOffer = offerRepository.save(offer);
-				newOffer.getUser().cleanPrivateDate();
+				newOffer.getUser().clearPrivateData();
 				return new ResponseEntity<Offer>(newOffer, HttpStatus.OK);
 			}
 			return new ResponseEntity<List<String>>(newOffer.getValidationErrors(), HttpStatus.BAD_REQUEST);
@@ -71,7 +72,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 	public ResponseEntity<?> delete(Long id) {
 		if (isNotNull(id)) {
 			Offer offerToDelete = offerRepository.findOne(id);
-			if (offerToDelete.getUser().getId() == getCurrentUser().getId()) {
+			if (isNotNull(offerToDelete) && offerToDelete.getUser().getId() == getCurrentUser().getId()) {
 				offerToDelete.setOfferStatus(OfferStatus.Closed);
 				offerRepository.save(offerToDelete);
 				return new ResponseEntity<String>("Oferta encerrada.", HttpStatus.OK);
@@ -87,7 +88,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 		if (isNotNull(id)) {
 			Offer offerDB = offerRepository.findOne(id);
 			if (isNotNull(offerDB)) {
-				offerDB.getUser().cleanPrivateDate();
+				offerDB.getUser().clearPrivateData();
 				return new ResponseEntity<Offer>(offerDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Nível de crescimento não encontrado.", HttpStatus.NOT_FOUND);
@@ -100,7 +101,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 		if (isNotNull(id)) {
 			Iterable<Offer> offerListDB = offerRepository.findAllByUser(id);
 			if (isNotNull(offerListDB)) {
-				offerListDB.forEach(offer -> offer.getUser().cleanPrivateDate());
+				offerListDB.forEach(offer -> offer.getUser().clearPrivateData());
 				return new ResponseEntity<Iterable<Offer>>(offerListDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Usuário não encontrado.", HttpStatus.NOT_FOUND);
@@ -114,7 +115,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 			FlowerShop flowerShopDB = flowerShopRepository.findOne(id);
 			if (isNotNull(flowerShopDB)) {
 				Iterable<Offer> offerListDB = offerRepository.findAllByFlowerShop(flowerShopDB.getId());
-				offerListDB.forEach(offer -> offer.getUser().cleanPrivateDate());
+				offerListDB.forEach(offer -> offer.getUser().clearPrivateData());
 				return new ResponseEntity<Iterable<Offer>>(offerListDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Floricultura não encontrada.", HttpStatus.NOT_FOUND);
@@ -128,7 +129,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 			Species speciesDB = speciesRepository.findOne(id);
 			if (isNotNull(speciesDB)) {
 				Iterable<Offer> offerListDB = offerRepository.findAllBySpecies(speciesDB.getId());
-				offerListDB.forEach(offer -> offer.getUser().cleanPrivateDate());
+				offerListDB.forEach(offer -> offer.getUser().clearPrivateData());
 				return new ResponseEntity<Iterable<Offer>>(offerListDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Espécie não encontrada.", HttpStatus.NOT_FOUND);
@@ -142,7 +143,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 			State stateDB = stateRepository.findOne(id);
 			if (isNotNull(stateDB)) {
 				Iterable<Offer> offerListDB = offerRepository.findAllByState(stateDB.getId());
-				offerListDB.forEach(offer -> offer.getUser().cleanPrivateDate());
+				offerListDB.forEach(offer -> offer.getUser().clearPrivateData());
 				return new ResponseEntity<Iterable<Offer>>(offerListDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Estado não encontrado.", HttpStatus.NOT_FOUND);
@@ -156,7 +157,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 			City cityDB = cityRepository.findOne(id);
 			if (isNotNull(cityDB)) {
 				Iterable<Offer> offerListDB = offerRepository.findAllByCity(cityDB.getId());
-				offerListDB.forEach(offer -> offer.getUser().cleanPrivateDate());
+				offerListDB.forEach(offer -> offer.getUser().clearPrivateData());
 				return new ResponseEntity<Iterable<Offer>>(offerListDB, HttpStatus.OK);
 			}
 			return new ResponseEntity<String>("Cidade não encontrada.", HttpStatus.NOT_FOUND);
@@ -172,8 +173,7 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 				OfferComment newOfferComment = new OfferComment(text, getCurrentUser(), offerDB);
 				if(newOfferComment.isValid()) {
 					newOfferComment = offerCommentRepository.save(newOfferComment);
-					newOfferComment.getUser().cleanPrivateDate();
-					newOfferComment.getOffer().getUser().cleanPassword();
+					newOfferComment.getUser().clearPrivateData();
 					return new ResponseEntity<OfferComment>(newOfferComment, HttpStatus.OK);
 				}
 				return new ResponseEntity<List<String>>(offerDB.getValidationErrors(), HttpStatus.BAD_REQUEST);
@@ -186,8 +186,12 @@ public class OfferServiceImpl extends IsHelper implements OfferService {
 	@Override
 	public ResponseEntity<?> deleteComment(Long id) {
 		if (isNotNull(id)) {
-			offerCommentRepository.delete(id);
-			return new ResponseEntity<String>("Comentário deletado", HttpStatus.OK);
+			OfferComment offerCommentDB  = offerCommentRepository.findOne(id);
+			if(offerCommentDB.getUser().getId() == getCurrentUser().getId()) {
+				offerCommentRepository.delete(id);
+				return new ResponseEntity<String>("Comentário deletado", HttpStatus.OK);
+			}
+			return new ResponseEntity<String>("Comentário não pertence ao usuário atual.", HttpStatus.UNAUTHORIZED);
 		}
 		return new ResponseEntity<String>("ID não pode ser nulo.", HttpStatus.BAD_REQUEST);
 	}
